use workers

+++ src/server.ts
@@ -1,3 +1,5 @@
+import dotenv from 'dotenv'
+dotenv.config()
 import { log } from './utils/log'
 import { messageFns } from './messages'
 import { websocketServer } from './websocket/websocket-server'
@@ -9,6 +11,8 @@ import {
 } from './metrics/metrics'
 import './http/http-server'
 import { DataChannelRepo } from './data/data-channel-repo'
+import cluster from 'cluster'
+import { spawn } from './spawn'
 
 const collectDefaultMetrics = prometheusClient.collectDefaultMetrics
 collectDefaultMetrics()
@@ -25,49 +29,51 @@ const server = async () => {
     redis.publish
   )
 
-  wss.on('connection', (ws) => {
-    log.debug({ event: `ClientConnected`, clients: wss.clients.size })
-    connectedClientsGauge.set(wss.clients.size)
+  if (cluster.isPrimary) {
+    spawn()
+  } else {
+    wss.on('connection', (ws) => {
+      log.debug({ event: `ClientConnected`, clients: wss.clients.size })
+      connectedClientsGauge.set(wss.clients.size)
 
-    ws.isAlive = true
-
-    ws.on('pong', () => {
       ws.isAlive = true
-    })
 
-    ws.onmessage = async (event) => {
-      try {
-        incomingMessageCounter.inc()
-        const result = await handleIncomingMessage(ws, event.data.toString())
+      ws.on('pong', () => {
+        ws.isAlive = true
+      })
 
-        if (result.isErr()) {
-          const error = result.error
-          if (error.message === 'write EPIPE') return
+      ws.onmessage = (event) => {
+        try {
+          incomingMessageCounter.inc()
+          handleIncomingMessage(ws, event.data.toString()).mapErr((error) => {
+            if (
+              [
+                'write EPIPE',
+                'Cannot call write after a stream was destroyed',
+              ].includes(error.message)
+            )
+              return
+            log.error(error)
+          })
+        } catch (error) {
           log.error(error)
         }
-
-        // .((err) => {
-        //   if (err.message === 'write EPIPE') return
-        //   return log.error(err)
-        // })
-      } catch (error) {
-        console.error(error)
       }
-    }
 
-    ws.onerror = (event) => {
-      log.error(event.error)
-    }
+      ws.onerror = (event) => {
+        log.error(event.error)
+      }
 
-    ws.onclose = () => {
-      connectedClientsGauge.set(wss.clients.size)
-      log.trace({
-        event: 'ClientDisconnected',
-        clients: wss.clients.size,
-      })
-      dataChannelRepo.remove(ws)
-    }
-  })
+      ws.onclose = () => {
+        connectedClientsGauge.set(wss.clients.size)
+        log.trace({
+          event: 'ClientDisconnected',
+          clients: wss.clients.size,
+        })
+        dataChannelRepo.remove(ws)
+      }
+    })
+  }
 }
 
 const runServer = async () => {
